made filename a command line argument for the database dump

+++ database/export-database.js
@@ -1,14 +1,24 @@
 /**
- * Script that export the Strapi database from the database Docker container that is running.
+ * Script that export the Strapi database from the database Docker container
+ * that is running.
  *
- * Stores the database dump in the local directory as a zip file that can be later used for importing to another
- * database instance.
+ * Stores the database dump in the local directory as a zip file that can be
+ * later used for importing to another database instance.
  */
 const { execShellCommand }  = require('./execShellCommand');
 const path = require("path");
+const ARGV_INDEX_FILE_PATH = 2;
+const ARGV_REQUIRED_LENGTH = 3;
 
 // Anonymous method that is directly called to allow for async/await usage
 (async () => {
+    // Verify a filename was provided.
+    if(process.argv.length < ARGV_REQUIRED_LENGTH){
+        console.error(`Please provide a filename name for the database export.`);
+        return;
+    }
+    const fileName = process.argv[ARGV_INDEX_FILE_PATH];
+
     try {
         // Delete any dump directory on the container
         await execShellCommand(`docker exec database rm -rf /dump`);
@@ -19,7 +29,7 @@ const path = require("path");
         // Compress the database dump into one zip file
         await execShellCommand(`docker exec database zip -r /dump.zip /dump`);
         // Copy the dump zip file over to the local directory
-        await execShellCommand(`docker cp database:/dump.zip ${path.resolve(__dirname, `dump.zip`)}`);
+        await execShellCommand(`docker cp database:/dump.zip ${path.resolve(__dirname, fileName)}`);
         // Log successful execution
         console.log(`Export successful!`);
     } catch(e){
